% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefsplot.R
\name{gg_coefsplot}
\alias{gg_coefsplot}
\title{Caterpillar plots of regression coefficients from a boral model}
\usage{
gg_coefsplot(model, X.labels = NULL)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}

\item{X.labels}{An optional vector of text strings to use as labels
for the X variables (graph facet labels).
If a named vector is provided (safest), the names will
be used to match labels to the X variable names in the fitted model.
If \code{NULL} (default), X variable names are taken from the model.}
}
\description{
Caterpillar plots of regression coefficients from a boral model
}
