% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefsplot.R
\name{gg_coefsplot_data}
\alias{gg_coefsplot_data}
\title{Prepare data for caterpillar plots of boral regression coefficients}
\usage{
gg_coefsplot_data(model)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}
}
\value{
A data frame for use with ggplot.
}
\description{
Prepare data for caterpillar plots of boral regression coefficients
}
