% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvsplot.R
\name{gg_lvsplot}
\alias{gg_lvsplot}
\title{Plot the latent variables from a boral model}
\usage{
gg_lvsplot(model, scaling = 1, lvs = c(1, 2))
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}

\item{scaling}{A numeric value between 0.8 and 1.2 that controls the
relative scaling of latent variables (e.g. sites) and latent variable
coefficients (e.g. species).}

\item{lvs}{Integer indices of the latent variables to plot.}
}
\value{
A ggplot object for the latent variable graph.
}
\description{
This is a ggplot version of the boral function \code{\link[boral]{lvsplot}}
to draw an ordination-style graph of latent variables (sites or objects)
and their corresponding coefficients (species or attributes).
}
\note{
Presently this function only graphs two latent variables.
}
