% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvsplot.R
\name{gg_lvsplot_data}
\alias{gg_lvsplot_data}
\title{Prepare data to plot latent variables from a boral model}
\usage{
gg_lvsplot_data(model, scaling = 1)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}

\item{scaling}{A numeric value between 0.8 and 1.2 that controls the
relative scaling of latent variables (e.g. sites) and latent variable
coefficients (e.g. species).}
}
\value{
A data frame with \code{n} column(s) for the values of the \code{n}
  latent variables; a column \code{var} specifying whether a row is for
  latent variable (\code{'lv'}) or coefficient (\code{'lvcoef'}) values;
  and a column \code{label} giving object (e.g. site) name for 'lv' rows
  and attribute (e.g. species) name for 'lvcoef' rows.
}
\description{
This function takes a fitted boral model and returns a data frame to use
with ggplot for an ordination-style bi-plot of latent variables and
coefficients. It is used by function \code{\link{gg_lvsplot}} but you
can also use it directly.
}
\examples{
\dontrun{
dat.gg <- gg_lvsplot_data(mymodel)

# Draw 'manually' with ggplot as an alternative to using
# the gg_lvsplot function
library(ggplot2)
library(ggrepel)

ggplot(data = dat.gg, aes(x = lv1, y = lv2, colour = var)) +
  geom_point() +
  geom_text_repel(aes(label = label)) +
  theme(legend.position = "none")
}

}
\seealso{
\code{\link{gg_lvsplot}}
}
