% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_plots.R
\name{gg_varpart}
\alias{gg_varpart}
\title{Plot variance components as estimated by calc.varpart}
\usage{
gg_varpart(model)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}
}
\value{
A ggplot object.
}
\description{
Plot variance components as estimated by calc.varpart
}
\examples{
#' library(boral)
library(ggboral)

data(spider, package = "mvabund")
y <- spider$abun
X <- scale(spider$x)

# Warning - these settings are only to make the example run quickly.
# Don't use them for a real analysis!
example.control <- list(n.burnin = 10, n.iteration = 100, n.thin = 1)

# Note that we specify save.model = TRUE when calling boral
#
spiderfit_nb <- boral(y, X,
                      family = "negative.binomial",
                      num.lv = 2,
                      row.eff = "fixed",
                      mcmc.control = example.control,
                      save.model = TRUE)

gg_varpart(spiderfit_nb) + theme_bw()

}
\seealso{
\code{\link[boral]{calc.varpart}}  \code{\link{gg_varpart_data}}
}
