% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_plots.R
\name{gg_varpart_data}
\alias{gg_varpart_data}
\title{Prepare data to plot variance components for a boral model}
\usage{
gg_varpart_data(model)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}
}
\value{
A data frame to use with ggplot that has the following columns:
  \describe{
    \item{varpart}{Variance component: one of X (explanatory variables),
      lv (latent variables), row (row effects).}
    \item{value}{Estimated proportion of variance.}
    \item{label}{Name of the response (e.g. species).}
  }
}
\description{
This function calls the boral function \code{\link[boral]{calc.varpart}} to
estimate the proportion of variance accounted for by explanatory variables,
latent variables and any row effects. It returns the results as a data frame
suitable for plotting with ggplot. It is used by \code{\link{gg_varpart}} but
you can also use it directly if you are constructing plots 'manually'.
}
\details{
This function requires that you fitted the model with explanatory variables
and specified \code{save.model = TRUE} when calling boral so that MCMC samples
were stored.
}
\examples{
#' library(boral)
library(ggboral)

data(spider, package = "mvabund")
y <- spider$abun
X <- scale(spider$x)

# Warning - these settings are only to make the example run quickly.
# Don't use them for a real analysis!
example.control <- list(n.burnin = 10, n.iteration = 100, n.thin = 1)

# Note that we specify save.model = TRUE when calling boral
#
spiderfit_nb <- boral(y, X,
                      family = "negative.binomial",
                      num.lv = 2,
                      row.eff = "fixed",
                      mcmc.control = example.control,
                      save.model = TRUE)

dat.varpart <- gg_varpart_data(spiderfit_nb)
head(dat.varpart)

}
\seealso{
\code{\link[boral]{calc.varpart}} \code{\link{gg_varpart}}
}
