% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nonzeroness}
\alias{nonzeroness}
\title{Describe where intervals lie in relation to zero}
\usage{
nonzeroness(lowers, uppers)
}
\arguments{
\item{lowers}{A vector of lower bounds.}

\item{uppers}{A vector of upper bounds the same length as \code{lowers}.}
}
\value{
A vector of nonzeroness values.
}
\description{
This is a helper function used by \code{\link{gg_coefsplot}} to colour
intervals based on where they lie in relation to zero. It returns a
value of 0.0 for intervals centred on zero; values between
0.0 and 1.0 for intervals containing zero away from their centre;
and 1.0 for intervals that exclude zero. This assumes that the intervals
represent symmetrical distributions, which often will not be the case.
}
