% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_num_lvs}
\alias{get_num_lvs}
\title{Gets the number of latent variables from a boral model object}
\usage{
get_num_lvs(model)
}
\arguments{
\item{model}{A fitted \code{boral} object.}
}
\value{
Number of latent variables.
}
\description{
Starting with version 1.7 of the boral package, the number of latent
variables is specified via the \code{lv.control} list argument to the
\code{boral} function. Prior to this it was specified using the \code{num.lv}
argument. This function detects which has been used.
}
