% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvsplot.R
\name{gg_lvsplot}
\alias{gg_lvsplot}
\title{Plot the latent variables from a boral model}
\usage{
gg_lvsplot(
  model,
  scaling = 1,
  lvs = c(1, 2),
  include = c("both", "objects", "attributes")
)
}
\arguments{
\item{model}{A boral model fitted with one or more latent variables.}

\item{scaling}{A numeric value between 0.8 and 1.2 that controls the
relative scaling of latent variables (e.g. sites) and latent variable
coefficients (e.g. species).}

\item{lvs}{Integer indices of the latent variables to plot.}

\item{include}{One of the following: \code{'both'} (default) to display both
objects and attributes; \code{'objects'} to display objects only; or
\code{'attributes'} to display attributes only.}
}
\value{
A ggplot object for the latent variable graph.
}
\description{
This is a ggplot version of the boral function \code{\link[boral]{lvsplot}}
to draw an ordination-style graph of latent variables (sites or objects)
and their corresponding coefficients (species or attributes).
}
\note{
Presently this function only graphs two latent variables.
}
\examples{
# This example is based on that for the lvsplot function
# in the boral package and requires an example dataset from the
# mvabund package.
library(boral)
library(ggboral)

data(spider, package = "mvabund")
y <- spider$abun

# Warning - these settings are only to make the example run quickly.
# Don't use them for a real analysis!
example.control <- list(n.burnin = 10, n.iteration = 100, n.thin = 1)

spiderfit_nb <- boral(y, family = "negative.binomial",
                      lv.control = list(num.lv = 2),
                      row.eff = "fixed",
                      mcmc.control = example.control)

gg_lvsplot(spiderfit_nb)

# Since the function returns a ggplot object you can tweak it further.
# Here we add labels and a white background.
last_plot() +
  labs(x = "Latent variable 1", y = "Latent variable 2",
       title = "Example unconstrained model of spider abundance") +

  theme_bw()

# Only display sites (objects) and not species (attributes)
#
gg_lvsplot(spiderfit_nb, include = "obj") +
  labs(x = "Latent variable 1", y = "Latent variable 2",
       title = "Site relationships from unconstrained model") +

  theme_bw()

}
\seealso{
\code{\link{gg_lvsplot_data}}
}
